refactor: Player의 name필드를 Name클래스로 변경
